requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
chromadb
langchain-core 
langchain-community
langchain-openai
langchain-chroma